Live coding session

+++ server/src/main/java/com/rofish/server/api/users/repositories/UserRepository.java
@@ -4,7 +4,9 @@ import com.rofish.server.api.users.models.User;
 import jakarta.validation.constraints.NotNull;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
+
     boolean existsByEmail(@NotNull String email);
+
     User getUserByEmail(String email);
 }
